fix(pages): update SPA fallback handling and restructure imports
- move SPA fallback functions to a dedicated module for better organization
- adjust import paths across multiple files to reflect new structure
- ensure SPA fallback logic is correctly referenced in all relevant components
